packages/python/pyproject.toml
[tool.poetry]
name = "x402-hpke"
version = "0.1.0"
description = "Provider-agnostic HPKE envelope library for x402 (Python)"
authors = ["Tim Cotten <tcotten@scrypted.ai>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "x402_hpke" }]

[tool.poetry.dependencies]
python = "^3.10"
PyNaCl = "^1.5.0"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"